password encryption, bean validation

+++ src/main/java/com/torrens/musicshop/repos/CommentRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CommentRepo extends JpaRepository<Comment, Integer> {
+public interface CommentRepo extends JpaRepository<Comment, Long> {
 
-    List<Comment> findByInstrumentId(Integer instrumentId);
+    List<Comment> findByInstrumentId(Long instrumentId);
 }
